Move test to same package as script
Moving the test file to `evals` directory as an annoying workaround to
allow importing `eval_gs_v1` which is both a command line script as well
as treats the evals dir as a package from which to import sibling
scripts as modules.
