+++ setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='marqo-gcl',
+    version='0.1',
+    packages=find_packages(),
+)
