+++ tox.ini
@@ -5,5 +5,4 @@ skipsdist = True
 
 [testenv]
 deps = -rrequirements-dev.txt
-setenv = PYTHONPATH = {toxinidir}
-commands = pytest tests/
+commands = pytest
